rewrote Project and User thingies
