+++ src/main/kotlin/ru/quipy/api/ProjectAggregateDomainEvents.kt
@@ -17,23 +17,23 @@ const val STATUS_DELETED_EVENT = "STATUS_DELETED_EVENT"
 @DomainEvent(name = PROJECT_CREATED_EVENT)
 class ProjectCreatedEvent(
         val projectId: UUID,
-        val projectName: String,
-        val creatorId: UUID,
+        var tasksAndStatusAggregateId: UUID,
+        val projectName: String
         createdAt: Long = System.currentTimeMillis(),
 ) : Event<ProjectAggregate>(
         name = PROJECT_CREATED_EVENT,
         createdAt = createdAt,
 )
 
-@DomainEvent(name = PROJECT_UPDATED_EVENT)
-class ProjectUpdatedEvent(
-        val projectId: UUID,
-        val projectName: String,
-        createdAt: Long = System.currentTimeMillis(),
-) : Event<ProjectAggregate>(
-        name = PROJECT_UPDATED_EVENT,
-        createdAt = createdAt,
-)
+//@DomainEvent(name = PROJECT_UPDATED_EVENT)
+//class ProjectUpdatedEvent(
+//        val projectId: UUID,
+//        val projectName: String,
+//        createdAt: Long = System.currentTimeMillis(),
+//) : Event<ProjectAggregate>(
+//        name = PROJECT_UPDATED_EVENT,
+//        createdAt = createdAt,
+//)
 
 @DomainEvent(name = PARTICIPANT_ADDED_EVENT)
 class ParticipantAddedEvent(
@@ -43,39 +43,4 @@ class ParticipantAddedEvent(
 ) : Event<ProjectAggregate>(
         name = PARTICIPANT_ADDED_EVENT,
         createdAt = createdAt,
-)
-
-@DomainEvent(name = LEAVE_PROJECT_EVENT)
-class LeaveProjectEvent(
-        val projectId: UUID,
-        val userId: UUID,
-        createdAt: Long = System.currentTimeMillis(),
-) : Event<ProjectAggregate>(
-        name = LEAVE_PROJECT_EVENT,
-        createdAt = createdAt,
-)
-
-@DomainEvent(name = STATUS_CREATED_EVENT)
-class StatusCreatedEvent(
-        val projectId: UUID,
-        val statusId: UUID,
-        val statusName: String,
-        val color: String,
-        createdAt: Long = System.currentTimeMillis(),
-) : Event<ProjectAggregate>(
-        name = STATUS_CREATED_EVENT,
-        createdAt = createdAt,
-)
-
-
-//color = if (Color.getColor(event.color) != null) Color.getColor(event.color) else throw IllegalArgumentException("Incorrect color type.")
-
-@DomainEvent(name = STATUS_DELETED_EVENT)
-class StatusDeletedEvent(
-        val projectId: UUID,
-        val statusId: UUID,
-        createdAt: Long = System.currentTimeMillis(),
-) : Event<ProjectAggregate>(
-        name = STATUS_DELETED_EVENT,
-        createdAt = createdAt
 )
