+++ src/main/kotlin/ru/quipy/config/EventSourcingLibConfiguration.kt
@@ -8,9 +8,9 @@ import ru.quipy.api.ProjectAggregate
 import ru.quipy.api.TaskAndStatusAggregate
 import ru.quipy.api.UserAggregate
 import ru.quipy.core.EventSourcingServiceFactory
-import ru.quipy.logic.ProjectAggregateState
-import ru.quipy.logic.TaskAndStatusAggregateState
-import ru.quipy.logic.UserAggregateState
+import ru.quipy.logic.state.ProjectAggregateState
+import ru.quipy.logic.state.TaskAndStatusAggregateState
+import ru.quipy.logic.state.UserAggregateState
 import ru.quipy.projections.AnnotationBasedProjectEventsSubscriber
 import ru.quipy.streams.AggregateEventStreamManager
 import ru.quipy.streams.AggregateSubscriptionsManager
