+++ src/main/kotlin/ru/quipy/controller/UserController.kt
@@ -3,7 +3,10 @@ package ru.quipy.controller
 import org.springframework.web.bind.annotation.*
 import ru.quipy.api.*
 import ru.quipy.core.EventSourcingService
+import ru.quipy.dto.user.CreateUserDto
 import ru.quipy.logic.*
+import ru.quipy.logic.command.createUser
+import ru.quipy.logic.state.UserAggregateState
 import java.util.*
 
 @RestController
@@ -13,13 +16,14 @@ class UserController(
 ) {
 
     @PostMapping("")
-    fun createUser(@RequestBody body: CreateUserRequest): UserCreatedEvent {
+    fun createUser(@RequestBody body: CreateUserDto): UserCreatedEvent {
+//        TODO: Check body.nickname to be unique
+
         return usersEsService.create { it.createUser(UUID.randomUUID(), body.uname, body.nickname, body.password) }
     }
-}
 
-data class CreateUserRequest (
-        val uname: String,
-        val nickname: String,
-        val password: String
-)
+    @GetMapping("/{id}")
+    fun getUser(@PathVariable id: UUID) : UserAggregateState? {
+        return usersEsService.getState(id)
+    }
+}
