+++ src/main/kotlin/ru/quipy/dto/project/AddParticipentDto.kt
@@ -1,0 +1,8 @@
+package ru.quipy.dto.project
+
+import java.util.*
+
+data class AddParticipantDto(
+        val userId: UUID
+)
+
