+++ src/main/kotlin/ru/quipy/dto/project/CreateProjectDto.kt
@@ -1,0 +1,8 @@
+package ru.quipy.dto.project
+
+import java.util.*
+
+data class CreateProjectDto(
+        val projectName: String,
+        val creatorId: UUID
+)
