+++ src/main/kotlin/ru/quipy/dto/taskandstatus/AddExecutorDto.kt
@@ -1,0 +1,8 @@
+package ru.quipy.dto.taskandstatus
+
+import java.util.*
+
+data class AddExecutorDto(
+        val userId: UUID
+)
+
