+++ src/main/kotlin/ru/quipy/dto/taskandstatus/CreateStatusDto.kt
@@ -1,0 +1,6 @@
+package ru.quipy.dto.taskandstatus
+
+data class CreateStatusDto (
+        val statusName: String,
+        val color: String
+)
