+++ src/main/kotlin/ru/quipy/dto/taskandstatus/CreateTaskDto.kt
@@ -1,0 +1,6 @@
+package ru.quipy.dto.taskandstatus
+
+data class CreateTaskDto(
+        val name: String,
+        val description: String
+)
