+++ src/main/kotlin/ru/quipy/dto/taskandstatus/UpdateTaskDto.kt
@@ -1,0 +1,6 @@
+package ru.quipy.dto.taskandstatus
+
+data class UpdateTaskDto(
+        val name: String,
+        val description: String
+)
