+++ src/main/kotlin/ru/quipy/dto/user/CreateUserDto.kt
@@ -1,0 +1,7 @@
+package ru.quipy.dto.user
+
+data class CreateUserDto(
+        val nickname: String,
+        val password: String,
+        val uname: String,
+)
