+++ src/main/kotlin/ru/quipy/entity/StatusEntity.kt
@@ -1,0 +1,10 @@
+package ru.quipy.entity
+
+import java.util.*
+
+data class StatusEntity(
+        val id: UUID,
+        val name: String,
+        val color: String
+)
+
