+++ src/main/kotlin/ru/quipy/enum/ColorEnum.kt
@@ -1,0 +1,13 @@
+package ru.quipy.enum
+
+enum class ColorEnum {
+    RED,
+    PINK,
+    ORANGE,
+    YELLOW,
+    GREEN,
+    BLUE,
+    PURPLE,
+    BROWN,
+    GREY
+}
