+++ src/main/kotlin/ru/quipy/logic/command/UserAggregateCommands.kt
@@ -1,6 +1,7 @@
-package ru.quipy.logic
+package ru.quipy.logic.command
 
 import ru.quipy.api.UserCreatedEvent
+import ru.quipy.logic.state.UserAggregateState
 import java.util.*
 
 fun UserAggregateState.createUser(id: UUID, userName: String, nickname: String, password: String): UserCreatedEvent {
